+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [[5.0.12](https://github.com/multiversx/mx-sdk-dapp/pull/1541)] - 2025-08-01
 
+- [Added shadow dom test utils](https://github.com/multiversx/mx-sdk-dapp/pull/1542)
 - [Removed loginInfo unused members](http://github.com/multiversx/mx-sdk-dapp/pull/1543)
 - [Fixed sign transactions error is not re-thrown](http://github.com/multiversx/mx-sdk-dapp/pull/1540)
 - [Fixed issue with account nonce](http://github.com/multiversx/mx-sdk-dapp/pull/1539)